style updated

+++ StudyWebsite/attachments/views.py
@@ -39,36 +39,41 @@ def add_attachment_view(request:HttpRequest,group_id):
     return render(request, "main/not_allowed.html") # no permission page
 
 
-def delete_attachment_view(request:HttpRequest,attachment_id ,group_id):
-  
-  group=StudyGroup.objects.get(pk=group_id)
+def delete_attachment_view(request: HttpRequest, attachment_id, group_id):
+    try:
+        attachment = Attachment.objects.get(pk=attachment_id)
+        group = StudyGroup.objects.get(pk=group_id)
 
-  try:
-    attachment=Attachment.objects.get(pk=attachment_id)
-    attachment.delete()
-  except Attachment.DoesNotExist:
-    return render(request, "main/not_found.html")
-  except Exception as e:
-    print(e)
-  return redirect("attachments:all_attachment_view",group_id=group.id)
-
-
-
-def update_attachment_view(request:HttpRequest,attachment_id ,group_id,user_id):
-      
-      # if not request.user.id == user_id or not request.user.is_superuser:
-      #   return render(request, "main/not_allwoed.html")
-      
-      group=StudyGroup.objects.get(pk=group_id)
-      attachment = Attachment.objects.get(pk=attachment_id)
-
-      if request.method == "POST":
-          try:
-            attachment.title = request.POST["title"]
-            attachment.media = request.FILES.get("media", attachment.media)
-    
-            attachment.save()
+        # Check if the current user uploaded the attachment
+        if request.user == attachment.uploaded_by:
+            attachment.delete()
             return redirect("attachments:all_attachment_view", group_id=group.id)
-          except Exception as e:
-              print(e)
-      return render(request, 'main/attachment.html',{"attachment":attachment,"group":group})
+        else:
+            return render(request, "main/not_allowed.html")  # no permission page
+
+    except Attachment.DoesNotExist:
+        return render(request, "main/not_found.html")
+    except Exception as e:
+        print(e)
+        # Handle other exceptions if necessary
+
+
+
+
+def update_attachment_view(request: HttpRequest, attachment_id, group_id):
+    group = StudyGroup.objects.get(pk=group_id)
+    attachment = Attachment.objects.get(pk=attachment_id)
+
+    # Check if the current user uploaded the attachment
+    if request.user == attachment.uploaded_by:
+        if request.method == "POST":
+            try:
+                attachment.title = request.POST["title"]
+                attachment.media = request.FILES.get("media", attachment.media)
+                attachment.save()
+                return redirect("attachments:all_attachment_view", group_id=group.id)
+            except Exception as e:
+                print(e)
+        return render(request, 'main/attachment.html', {"attachment": attachment, "group": group})
+    else:
+        return render(request, "main/not_allowed.html")  # no permission page
